Int32 for PID, Idle mode, monitoring switch, bug correction
Regulation uses now (U)Int32 instead of (U)Int16.
Angles are in micro-radians to improve accuracy (especially for the D
part of PID)
Regulation is reseted and motors remains idle when throttle is below
kReceiverIDLE

Monitoring: corrected frame start string

Monitoring is now on switch 4

+++ Sources/tasks/gFlightCompute.h
@@ -23,9 +23,9 @@ typedef enum {
 
 //Struct that contains the PID datas for each PID
 typedef struct {
-	Int16 aKp, aKi, aKd;
-	Int16 aPreviousError;
-	Int16 aIntegral;
+	Int32 aKp, aKi, aKd;
+	Int32 aPreviousError;
+	Int32 aIntegral;
 	UInt16 aPreviousTime;
 } PIDdata;
 PIDdata sPID [3];
